Use newtype wrapper for drawing `Job` with superconsole
