+++ crates/brioche-core/src/reporter/console.rs
@@ -9,6 +9,7 @@ use std::{
 
 use bstr::{BString, ByteSlice};
 use human_repr::HumanDuration as _;
+use joinery::JoinableIterator as _;
 use opentelemetry::trace::TracerProvider as _;
 use tracing_subscriber::{layer::SubscriberExt as _, util::SubscriberInitExt as _, Layer as _};
 
@@ -389,7 +390,7 @@ impl superconsole::Component for JobsComponent {
 
         let jobs_lines = job_list
             .map(|(_, job)| {
-                job.draw(
+                JobComponent(job).draw(
                     superconsole::Dimensions {
                         width: dimensions.width,
                         height: 1,
@@ -487,6 +488,130 @@ impl superconsole::Component for JobsComponent {
             .chain(jobs_lines.into_iter().flatten())
             .chain(summary_line)
             .collect();
+        Ok(lines)
+    }
+}
+
+struct JobComponent<'a>(&'a Job);
+
+impl<'a> superconsole::Component for JobComponent<'a> {
+    fn draw_unchecked(
+        &self,
+        _dimensions: superconsole::Dimensions,
+        _mode: superconsole::DrawMode,
+    ) -> anyhow::Result<superconsole::Lines> {
+        let job = &self.0;
+        let lines = match job {
+            Job::Download {
+                url,
+                progress_percent,
+            } => {
+                let message = match progress_percent {
+                    Some(100) => {
+                        format!("[100%] Downloaded {url}")
+                    }
+                    Some(progress_percent) => {
+                        format!("[{progress_percent:>3}%] Downloading {url}")
+                    }
+                    None => {
+                        format!("[???%] Downloading {url}")
+                    }
+                };
+                superconsole::Lines::from_iter([superconsole::Line::sanitized(&message)])
+            }
+            Job::Unarchive { progress_percent } => {
+                let message = if *progress_percent == 100 {
+                    "[100%] Unarchived".to_string()
+                } else {
+                    format!("[{progress_percent:>3}%] Unarchiving")
+                };
+                superconsole::Lines::from_iter([superconsole::Line::sanitized(&message)])
+            }
+            Job::Process {
+                packet_queue: _,
+                status,
+            } => {
+                let child_id = status
+                    .child_id()
+                    .map(|id| id.to_string())
+                    .unwrap_or_else(|| "?".to_string());
+                let elapsed = status.elapsed().human_duration();
+                let message = match status {
+                    ProcessStatus::Running { .. } => {
+                        format!("Process {child_id} [{elapsed}]")
+                    }
+                    ProcessStatus::Exited { status, .. } => {
+                        let status = status
+                            .as_ref()
+                            .and_then(|status| status.code())
+                            .map(|c| c.to_string())
+                            .unwrap_or_else(|| "?".to_string());
+                        format!("Process {child_id} [{elapsed} Exited {status}]")
+                    }
+                };
+
+                superconsole::Lines::from_iter(std::iter::once(superconsole::Line::sanitized(
+                    &message,
+                )))
+            }
+            Job::RegistryFetch {
+                complete_blobs,
+                total_blobs,
+                complete_recipes,
+                total_recipes,
+            } => {
+                let blob_percent = if *total_blobs > 0 {
+                    (*complete_blobs as f64 / *total_blobs as f64) * 100.0
+                } else {
+                    100.0
+                };
+                let recipe_percent = if *total_recipes > 0 {
+                    (*complete_recipes as f64 / *total_recipes as f64) * 100.0
+                } else {
+                    100.0
+                };
+                let total_percent = (recipe_percent * 0.2 + blob_percent * 0.8) as u8;
+                let verb = if job.is_complete() {
+                    "Fetched"
+                } else {
+                    "Fetching"
+                };
+                let fetched_blobs = if *total_blobs == 0 {
+                    None
+                } else if job.is_complete() {
+                    Some(format!(
+                        "{complete_blobs} blob{s}",
+                        s = if *complete_blobs == 1 { "" } else { "s" }
+                    ))
+                } else {
+                    Some(format!(
+                        "{complete_blobs} / {total_blobs} blob{s}",
+                        s = if *total_blobs == 1 { "" } else { "s" }
+                    ))
+                };
+                let fetched_recipes = if *total_recipes == 0 {
+                    None
+                } else if job.is_complete() {
+                    Some(format!(
+                        "{complete_recipes} recipe{s}",
+                        s = if *complete_recipes == 1 { "" } else { "s" }
+                    ))
+                } else {
+                    Some(format!(
+                        "{complete_recipes} / {total_recipes} recipe{s}",
+                        s = if *total_recipes == 1 { "" } else { "s" }
+                    ))
+                };
+                let fetching_message = [fetched_blobs, fetched_recipes]
+                    .into_iter()
+                    .flatten()
+                    .join_with(" + ");
+                let message =
+                    format!("[{total_percent:>3}%] {verb} {fetching_message} from registry",);
+                superconsole::Lines::from_iter([superconsole::Line::sanitized(&message)])
+            }
+        };
+
         Ok(lines)
     }
 }
